Add error case and refactor oauth controller

+++ app/controllers/tokens_controller.rb
@@ -1,4 +1,6 @@
 class TokensController < ApplicationController
+  before_action :validate_oauth_authorization, only: [:access_token]
+
   def request_token
     state = SecureRandom.hex
     Oauth.create(state: state)
@@ -10,22 +12,29 @@ class TokensController < ApplicationController
   end
 
   def access_token
+    access_token = LINKEDIN.auth_code.get_token(
+      params[:code],
+      redirect_uri: ENV['OAUTH_CALLBACK']
+    )
+    response = access_token.get(
+      'https://api.linkedin.com/v1/people/~?format=json'
+    )
+    json_response = JSON.parse(response.body)
+    user = User.find_or_create_by(uid: json_response["id"])
+    jwt = JWT.encode({uid: user.uid, exp: 1.day.from_now.to_i}, Rails.application.secrets.secret_key_base)
+    redirect_to ENV['ORIGIN'] + "?jwt=#{jwt}"
+  end
+
+  private
+
+  def validate_oauth_authorization
+    if params[:error]
+      redirect_to ENV['ORIGIN'] + '?error=login_error'
+    end
     oauth = Oauth.find_by_state(params[:state])
     if oauth.nil?
       redirect_to ENV['ORIGIN'] + '?error=login_error'
-    else
-      access_token = LINKEDIN.auth_code.get_token(
-        params[:code],
-        redirect_uri: ENV['OAUTH_CALLBACK']
-        )
-      response = access_token.get(
-        'https://api.linkedin.com/v1/people/~?format=json'
-        )
-      json_response = JSON.parse(response.body)
-      user = User.find_or_create_by(uid: json_response["id"])
-      jwt = JWT.encode({uid: user.uid, exp: 1.day.from_now.to_i}, Rails.application.secrets.secret_key_base)
-      oauth.destroy
-      redirect_to ENV['ORIGIN'] + "?jwt=#{jwt}"
     end
+    oauth.destroy
   end
 end
